CDR-1198 Apply changes from review and use Integer based properties

+++ response-dto/src/main/java/org/ehrbase/openehr/sdk/response/dto/ehrscape/QueryResultDto.java
@@ -26,8 +26,8 @@ public class QueryResultDto {
 
     private OffsetDateTime created;
     private String executedAQL;
-    private Long limit;
-    private Long offset;
+    private Integer limit;
+    private Integer offset;
     private List<List<String>> explain;
     private Map<String, String> variables;
     private List<ResultHolder> resultSet;
@@ -64,19 +64,19 @@ public class QueryResultDto {
         this.executedAQL = executedAQL;
     }
 
-    public @Nullable Long getLimit() {
+    public @Nullable Integer getLimit() {
         return limit;
     }
 
-    public void setLimit(Long limit) {
+    public void setLimit(Integer limit) {
         this.limit = limit;
     }
 
-    public @Nullable Long getOffset() {
+    public @Nullable Integer getOffset() {
         return offset;
     }
 
-    public void setOffset(Long offset) {
+    public void setOffset(Integer offset) {
         this.offset = offset;
     }
